Feat(Menu): A better menu:
Add settings in menu
Not making any changes for the moment
Tl;DR: UNSTABLE

+++ Sources/inc/HUDWindow.hpp
@@ -51,7 +51,7 @@ class	HUDWindow : public HUDActor {
 			int				colorR;
 			int				colorG;
 			int				colorB;
-			int				colorA;
+			float				colorA;
 			Characters		*toFollow;
 			int				isFading;
 			int				isTalk;
